Move query validator to different file
